Prototype: Support multiple worlds
