+++ prototype/src/js/constants.js
@@ -9,14 +9,6 @@ export const betaConstants = {
     max: 90
 };
 
-export const initialMapCenter = [39.96133, -75.15416];
-
-export const initialZoom = 21;
-
-export const defaultZoneRadius = 4; // meters
-
-export const zoneBuffer = 0.002; // kilometers
-
 export const cssOffsetPropertyName = '--offset';
 
 export const hideOrienterDelay = 700;
